Extract pipeline module from the monolith (split PR 8/N)
Moves the per-transcript pipeline stages (tfbs_finder, alignment
orchestration, and all Ensembl transcript/gene/regulatory metadata
retrieval) into tfbs_footprinter3/pipeline.py.

- tfbs_footprinter3/pipeline.py (new): tfbs_finder,
  retrieve_genome_aligned, alignment_tools, test_transcript_id,
  transfabulator, transcript_data_retrieve, gene_data_retrieve,
  retrieve_regulatory — each stage main() orchestrates per transcript.

- Ensembl REST calls use attribute access (ensembl.ensemblrest(...))
  at call time rather than a bare-name import at module load. That
  makes the HPC cache monkey-patch robust: patching
  tfbs_footprinter3.ensembl.ensemblrest transparently affects every
  pipeline-side caller.

- hpc/ensembl_cache.py::patch_tfbs_footprinter3 now rebinds BOTH
  tfbs_footprinter3.ensembl.ensemblrest (canonical location, used by
  pipeline callers) AND tfbs_footprinter3.tfbs_footprinter3.ensemblrest
  (the legacy re-export). Verified end-to-end: after applying the
  patch, a live call to the patched ensembl.ensemblrest returns
  species='homo_sapiens' for ENST00000361390.

- Drops the ## Py2 IUPAC alphabet shadow inside tfbs_finder (dead code
  since Biopython 1.78) as behavior-neutral cleanup.

Monolith: 853 -> 514 lines. 21 tests green, ruff clean.

Remaining in monolith (~500 lines):
- main() + get_args() + the re-export shim block. Last extraction
  candidate is cli.py covering main and get_args.
